requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# MongoDB
pymongo==4.5.0

# Redis (cache)
redis==5.0.1

# RPA e automação
selenium==4.15.2
playwright==1.40.0
pyautogui==0.9.54

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
httpx==0.25.1

# Data science (opcional)
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# IA local (Hugging Face - gratuito)
transformers==4.36.0
sentence-transformers==2.2.2
chromadb==0.4.22

# Testes
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# Logging e monitoramento
structlog==23.2.0

# Utilitários
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
